refactor: SearchPage.tsx の状態管理を useSearchPage フックに抽出する (#72)
